BugFix & some memory changes

+++ LEDMetr/src/fcs/interrupts.h
@@ -20,7 +20,7 @@ extern volatile int AveragedReadsCounter;
 extern volatile uint16_t AD_Data_Values[16];
 extern volatile uint8_t MultiplexerChannel;
 extern volatile int ChannelSwitchingCounter;
-extern volatile int NumberOfAveragedValues;
+extern volatile uint16_t NumberOfAveragedValues;
 extern volatile short int statusRS232;
 extern volatile short int pozRS232;
 extern volatile short int afterFirstQuote;
